Update build process
dev.sh starts the app in dev mode which mounts the local source files and (opposed to the old pattern) allows source "on the server" to be edited without reloading the app, e.g. because the source running "in the app" is located in the local file tree.
prod.sh starts the app in production mode which runs it more securly and with a configurable number of gunicorn workers (in the case of the api)